+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # News
 
-## v0.4.4-dev
+## v0.5.0-dev
 
+- **(breaking)** Rename all pre-defined tableaux to have a `t` prefix. e.g., `CNOT`→`tCNOT`, in order to distinguish them from "symbolic" operators like `sCNOT`.
+- **(breaking)** Rename `CliffordId` to `tId1` to match the naming style of `sId1`.
+- Implement convertors from symbolic operators to dense tableau operators: `CliffordOperator(sCNOT)` now gives `tCNOT` which acts equivalently to `sCNOT(1,2)`.
 - Implement `project[X|Y|Z]rand!` as a simpler interface to `project!` with automatic randomization of measurement phases.
 - Implement `sMX`/`sMY`/`sMZ` symbolic measurement operations that can be used with `apply!`. Use `projectrand!` internally.
 - Move `Register` out of `Experimental`. Used with `sMX`/etc to store classical bit results during circuit evolution.